refactor : ScoreStatus 클래스의 패키지를 baseball.domain 에서 baseball.domain.enums로 변경
패키지 변경에 따른 의존 관계 클래스의 import 문 변경
